network/requirements.txt
torch
torchvision
pytorch-lightning
vtk
matplotlib
numpy
Pillow
PyYAML
